use sample summary type
